Added MongoDB Template instead of Repository

+++ olms/src/main/java/com/se2/proj/olms/service/LectureService.java
@@ -12,36 +12,40 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 
-import com.se2.proj.olms.dao.LectureRepository;
 import com.se2.proj.olms.entities.LectureDocument;
 
 @Service
 public class LectureService {
- 
+
     @Autowired
-    private LectureRepository lectureRepository;
- 
+    private MongoTemplate mongoTemplate;
+
     @Value("${file.upload-dir}")
     private String uploadDir;
- 
+
     public List<LectureDocument> getLecturesByCourseId(String courseId) {
-        return lectureRepository.findByCourseId(courseId);
+        Query query = new Query();
+        query.addCriteria(Criteria.where("courseId").is(courseId));
+        return mongoTemplate.find(query, LectureDocument.class);
     }
- 
+
     public LectureDocument uploadLecture(String courseId, MultipartFile file) throws Exception {
         // Create course directory if it doesn't exist
         Path courseDir = Paths.get(uploadDir, courseId);
         Files.createDirectories(courseDir);
- 
+
         // Generate unique filename
         String originalFilename = StringUtils.cleanPath(file.getOriginalFilename());
         String filename = System.currentTimeMillis() + "_" + originalFilename;
         Path filePath = courseDir.resolve(filename);
- 
+
         // Save file to filesystem
         Files.copy(file.getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
- 
+
         // Create and save document in MongoDB
         LectureDocument lecture = new LectureDocument();
         lecture.setCourseId(courseId);
@@ -51,19 +55,33 @@ public class LectureService {
         lecture.setUploadDate(LocalDateTime.now());
         lecture.setFileType(file.getContentType());
         lecture.setFileSize(file.getSize());
- 
-        return lectureRepository.save(lecture);
+
+        mongoTemplate.save(lecture);
+        return lecture;
     }
- 
+
     public void deleteLecture(String id) throws Exception {
-        LectureDocument lecture = lectureRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Lecture not found"));
- 
+        // Retrieve lecture document
+        Query query = new Query();
+        query.addCriteria(Criteria.where("_id").is(id));
+        LectureDocument lecture = mongoTemplate.findOne(query, LectureDocument.class);
+        
+        if (lecture == null) {
+            throw new RuntimeException("Lecture not found");
+        }
+
         // Delete file from filesystem
         Path filePath = Paths.get(lecture.getFilePath());
         Files.deleteIfExists(filePath);
- 
+
         // Delete document from MongoDB
-        lectureRepository.deleteById(id);
+        mongoTemplate.remove(query, LectureDocument.class);
+    }
+
+    // Example of a custom query using MongoTemplate
+    public List<LectureDocument> findLecturesByCustomQuery(String courseId) {
+        Query query = new Query();
+        query.addCriteria(Criteria.where("courseId").is(courseId));
+        return mongoTemplate.find(query, LectureDocument.class);
     }
 }
